+++ AST_INC/AST/BasicBlock.hpp
@@ -7,23 +7,10 @@
 #include <map>
 #include <string>
 #include "ThreeAddressInstruction.hpp"
+#include "RegisterScope.hpp"
 
 namespace cs6300
 {
-struct RegMeta
-{
-    std::set<int> dead;
-    std::set<int> used;
-    std::set<int> alive;
-
-    void clear()
-    {
-        dead.clear();
-        used.clear();
-        alive.clear();
-    }
-};
-
 class BasicBlock
 {
 
@@ -34,12 +21,11 @@ class BasicBlock
         std::vector<std::shared_ptr<BasicBlock>> parents;
         int branchOn;
         std::string getLabel();
-        void printInstructions(bool sets=false);
 
         //Used for register allocation
-        RegMeta m;
+        cs6300::RegisterScope m;
         void initSets();
-        static RegMeta meta(ThreeAddressInstruction tal);
+        static cs6300::RegisterScope scope(ThreeAddressInstruction tal);
         void remap(std::map<int,int> m);
     private:
         std::string label;
